requirements.txt
flask
werkzeug
transformers
torch
accelerate
sentencepiece

easyocr
opencv-python

pillow
pytesseract

PyPDF2
python-docx
python-pptx